02_content_service/requirements.txt
fastapi>=0.111
uvicorn>=0.30
pydantic>=2.7
python-multipart>=0.0.9
pymupdf>=1.24
python-pptx>=1.0.2
Pillow>=10.0
scikit-learn>=1.5
reportlab>=4.0
pytest>=8.0
httpx>=0.27
